Desugar template parameters, so that they get stringified correctly.
In particular, `ElaboratedType` nodes cause a temporary suppression of scopes in strings. See `ElaboratedTypePolicyRAII`. This... makes it really seem like a mistake to use `ElaboratedType` at all for template parameters, but maybe there's some hidden benefit? At any rate, we _must_ remove the `ElaboratedType` when stringifying this in order to get a valid type name, and to do that, the only option I could figure out is to canonicalize the template parameters. Truthfully, that may be the right call anyway, I don't know.

This doesn't yet fix all the tests.

PiperOrigin-RevId: 466323870