De-`Elaborate` AST nodes even earlier.
My earlier "hacky" change removed this sugar in `ConvertType`, not `ConvertQualType`, but this is actually "too late": the error messages inside `ConvertQualType`, in that case, will be mangled by the `ElaboratedType` node, which removes contextual information. For example, instead of saying `struct Foo::NestedStruct`, it will say `NestedStruct`. Our tests (rightly) were for the clearer name.

In addition, calling back into `ConvertQualType` from `ConvertType` produces confusing error messages when the desugared type has the same string representation. You'd get errors like `Unsupported type Foo: Unsupported type Foo: <real error>`, because an `ElaboratedType(Foo(...))` and a `Foo(...)` might both stringify as simply `Foo`.

So the simplest fix all around is to desugar the `QualType`, not the `Type`. It took some research to figure out how you do this -- for example, do I return a `SplitQualType` or a `QualType`? Well you see, for template const propagation to work we need to look up the canonical type, which is only convenient from `QualType`, so `SplitQualType` is no good...

Anyway, this fixes the last failures in the linked bug.

PiperOrigin-RevId: 466324077