Make sure structs/unions that have non-unpin fields don't implement Copy
In the current code, Record::is_unpin() is defined as
  self.is_trivial_abi && !self.is_inheritable

but doesn't look into the Record's fields recursively. Therefore
for the following code

  struct TrivialButInheritable {
    int x;
  };
  struct StructWithInheritable final {
    TrivialButInheritable t;
  };

This makes StructWithInheritable implement Copy while TrivialButInheritable does not,
and therefore results in a compile error.

This change adds .is_inheritable field to the Field struct, and check the
field recursively when Record::is_unpin() is called.

PiperOrigin-RevId: 466962560