Implement bindings from unary `operator-` and `operator!` to `Neg` and `Not`.
Excluded are `operator+` and `operator~`, which have no Rust equivalent.

I considered mapping `operator~` to `Not` instead of `operator!`, which would match the bitwise negation semantic for integers, but mapping `!` to `!` seemed least confusing. The mapping for `operator<<`, used for stream insertion in C++, similarly maps directly to `Shl` in rust, preserving the symbol while not necessarily being idiomatic.

PiperOrigin-RevId: 468260624