Update record.size when the original size in C++ isn't aligned (while alignment is given)
This typically only happens when a record (struct or union) is an anonymous
declaration with a typedef name, and the outer typedef record has the alignment
attribute (but not the underlying record field).

We need to also stash the original unaligned size though because that is what
sizeof(record) returns, and the binding tries to produce static_asserts for
the size.

PiperOrigin-RevId: 468486613