Remove /dev/shm trick from the bazelrc
This was added as a performance optimization, but in some docker environments
(such as the one in our CI) shm can be as small as 64mbs, and that is just not
enough for LLVM. While we investigate our way forward in
https://github.com/bazelbuild/continuous-integration/issues/1379#issuecomment-1228442455
we remove the flag from the project-specific bazelrc. If this flag is cricital to contributors we recommend adding it to the system wide or local bazelrcs.

PiperOrigin-RevId: 470700736