Fix comment of DeclImporter::GetItemIdsInSourceOrder.
Before this CL, the comment would incorrectly state that
GetItemIdsInSourceOrder "assumes that the children decls have already
been imported". This statement was incorrect, because implementation of
GetItemIdsInSourceOrder calls Importer::GetDeclItem which "returns the
Item of a Decl, importing it first if necessary".

Other minor tweaks have also been applied to the comment.

This CL has no impact on the behavior of the code.

PiperOrigin-RevId: 471825880