Don't generate trait `impl`s for pointers.
Without `#pragma clang lifetime_elision` we won't infer Rust lifetime of
`this` and will attempt to represent it as a Rust pointer (rather than
as a Rust reference).  We should not generate trait `impl`s for such
pointers.  One major reason is that pointers are considered foreign
and violate Rust's orphan rule.

PiperOrigin-RevId: 471862340