Consistently handle explicit and implicit class template specializations.
If a template specialization is used in a type alias (or in a parameter
type, etc.) Crubit's IR includes a corresponding `Record` as a top-level
item.  OTOH, if a template specialization is *not* actually used, then
there is no need to include its `Record`.

Before this CL, the above was haphazardly achieved by checking
`ClassTemplateSpecializationDecl::isExplicitSpecialization` and skipping
explicit specializations, incorrectly assumming that such
specializations are not used in a type alias, etc.

After this CL, the above is achieved by skipping processing of
`ClassTemplateSpecializationDecl` in `GetCanonicalChildren` (and
therefore only processing such decls via
`Importer::ConvertTemplateSpecializationType`).

As a side-effect, this CL also fixes a bug, where we would assume that
all `ClassTemplateSpecializationDecl` have no `enclosing_namespace_id`,
but `GetCanonicalChildren` could still include such decls as children of
a namespace.  The CL adds coverage for this bug under
`test/templates/extern_definition`.

PiperOrigin-RevId: 473297905