Import `string_view` as if it were `final`.
Later, we should read this from a sidecar file, but for now, this should work. (It's a trivial change, once we actually have sidecar files, to use that instead.)

This is safe, because string_view has no padding, and with luck never will have padding. Also, nobody seriously inherits from it.

PiperOrigin-RevId: 473974990