Add more tests for reasons I remembered that it could be dangerous to pretend a type is `final`.
Really, honestly, `string_view` is the best case for this!

It has all the properties we'd want so that destroying it by value is -- maybe UB, but implementation-defined to be ok. It has no tail padding. It has no vtable. Its destructor is trivial, meaning in practice that it doesn't even do anything.

On top of that, people rarely, if ever, inherit from it. Even if they did inherit from it, it's predominantly passed by value anyway.

Anyway, this test should be reassuring, I hope.

PiperOrigin-RevId: 474234757