Rewrite constraint resolution with a different (hopefully correct) algorithm.
The previous algorithm considered *all* lifetime constraints (except those involving local or static) to be equality constraints. However, this is incorrect, because even if we cannot represent inequality constraints in the signature, inequality constraints between lifetimes that are not visible in the signature are still meaningful.

Moreover, the previous algorithm did not handle the case of unconstrained / outlived-by-static output lifetimes (which can be made 'static).

For example, consider:
```
'local1 >= 'local2
'param1 >= 'local2
```

The previous algorithm would incorrectly deduce that the function leaked a 'local1 lifetime through 'param1.

PiperOrigin-RevId: 474809494