Lift records behind typedefs into first-class (cached) items.
The previous algorithm was not sound: it would, as part of graph traversal, mark that a record would be successfully imported when it first encountered it, under the assumption that the import would be cached. However, typedefs specifically bypassed this caching. And so you could, if you tried to import a typedef _before_ the struct it imported, accidentally mark the struct as imported, without actually importing it, thus preventing future import and causing dangling DeclIDs. This change removes that, and should unblock future work to import more typedef declarations.

After this CL, all records which are successfully imported in some way have an Item, returned by the Record importer -- *not* the typedef importer. The typedef importer, if it sees sees anonymous structs, skips them.

Special handling of typedefs is, as a consequence, partially bled into the record importer. I've left a TODO in the place where this needs to be refactored in the future.

(This took me WAY too long to debug and write! Spooky action at a distance, yo.)

PiperOrigin-RevId: 475802057