rs_bindings_from_cc: avoid generating bindings for types of arguments of private methods.
Previously the check that verified visibility of a method came fairly late in the generation of bindings for a function.
This caused bindings to the type of some arguments of private methods to be generated, causing issues in C++ code during compilation.

The example in private_method.h failed to compile because crubit attempted to generate bindings to the specialization of Ptr<Inner> but Inner is not visibile to the binding code.

PiperOrigin-RevId: 476907408