Exclude <vector> from the supported STL.
For example, if a type is not default-constructible, then the use of `vector<type>` in the public API causes imports to fail, because `vector<type>::resize` can't be imported. Until we fix that, it's better not to include vector. Instead, the useful minimal test for template instantiation etc. is probably `string_view` and `string`, which is easier to get working.

PiperOrigin-RevId: 477030243