Suppress template function instantiations in user crates for STL functions.
This is the easiest way I can figure out to suppress functions like `std::vector<T>::resize()` (which breaks compilation if instantiated with a non-default-constructible T), given the two mentioned bugs.

This is not a complete loss -- in particular, we do still get instantiations inside the STL. So for example, we should still be able to get string, string_view, etc. working. But not `vector<T>` for any `T` just yet.

PiperOrigin-RevId: 477030556