Move graph traversal updates so that they are compatible with records being imported during record import.
This is pretty subtle, so I've split it out to a separate change, even though it doesn't happen right now, or at least not in any of our tests.

In particular, consider the following failure modes:

1. a record is imported during its own import. This causes the cache to be set to `nullopt` (since it's marked as already imported), and the insertion to fail. So we actually need to overwrite (or be even cleverer still).
2. a record is imported by its fields during its own import. Because in the old version of the code, it wasn't cached before it had the fields imported, this causes an infinite loop (and eventual stack overflow).

There may have been another failure mode I am forgetting.

In order to help debug the first kind of failure, where we foul the cache with `nullopt`, I've also changed it from `nullopt` to an unsupported item. This will never show up at all when things are working correctly, and is just to aid debugging.

PiperOrigin-RevId: 478945495