Add custom bindings glue code to the STL, starting with `std::string_view`.
This change makes `std::string_view` usable, in a hacky way, from Rust: you can convert a Rust `&'static str` or `&'static [u8]` to a C++ `std::string_view`, and vice versa.

This is not *sound*:

1. `string_view` doesn't necessarily obey Rust's aliasing rules. (This is common with all our bindings atm.)
2. `string_view` doesn't currently have lifetime params; we just kinda assume static for now. (This makes Rust code that takes in a string_view about as safe as C++ code that does, to be fair!)

Overall, not the worst. And the result is that you really can call APIs that take and return `string_view`. Well, probably.

### How does it work?

This reuses `extra_rs_srcs` so that the STL can have additional methods/trait impls. In theory the same mechanism should also work for providing extra sources to any target, but that requires more plumbing / aspect hints / whatever. At the moment this does not expose `extra_rs_srcs` to anything except the STL. But presumably that's an obvious next step, especially if we want to do similar things to support Abseil, and so on.

PiperOrigin-RevId: 478948595