Generate `operator<` bindings.
For now operator<=> is completely ignored, and even if both operator< and operator<=> are present we construct partial_cmp from lt (i.e. the operator< thunk).

Bindings are also not generated for operator< when the operands have different types, since that's a much more complicated case: implementing partial_cmp in Rust using lt would require transitivity of operator< and operator==, and we'd need to query salsa for them while avoiding cyclic queries.

PiperOrigin-RevId: 480291403