Add minimal support for opaque ADT bindings.
This CL adds support for generating bindings for `struct`, `enum`, or
`union`.  Currently the generated binding represents the type as an
opaque blob of bytes + requires that the wrapped type doesn't have a
custom `Drop` impl.  Additionally, no bindings are currently generated
for 1) the default constructor, 2) the copy constructor and the copy
assignment operator, nor 3) any member functions.

PiperOrigin-RevId: 490383109