s/ `String` / `Rc<str>` / in `NamespaceQualifier` and `ir::Identifier`.
This CL was motivated by trying to improve the efficiency of
`NamespaceQualifier`'s `Clone` implementation - cloning of `Rc<str>`
requires just increasing the ref-count and is therefore cheaper than
cloning of `String` which requires a heap allocation and copying the
string contents.

This CL also changes the type of `ir::Identifier` because this is
the representation of `ir::Namespace::name`.

`Rc<str>` instead of `String` is possible because these strings are
never mutated.

PiperOrigin-RevId: 490605310