Update nullability verification framework to adapt to the new diagnoser API in the clang dataflow framework
The new API (PHAB_REVIEW=https://reviews.llvm.org/D139868) allows diagnosers to take in `State` objects instead of `Environment` objects as argument. A `State` contains both an `Environment` and a `Lattice` element. This change will allow the pointer nullability diagnosis to use information from the dataflow lattice, which can be helpful for testing and debugging purposes. For example, the diagnoser can now access the nullability vectors computed for a given expression, which are stored in a map in the lattice.

PiperOrigin-RevId: 495350578