Handle double pointer dereferences
Consider the following code
```
void target(int * _Nullable * _Nonnull x){
    **x;
}
```
and its CFG representation:
```
void target(int * _Nullable * _Nonnull x)
 [B2 (ENTRY)]
   Succs (1): B1

 [B1]
   1: x
   2: [B1.1] (ImplicitCastExpr, LValueToRValue, int * _Nullable * _Nonnull)
   3: *[B1.2]
   4: [B1.3] (ImplicitCastExpr, LValueToRValue, int * _Nullable)
   5: *[B1.4]
   Preds (1): B2
   Succs (1): B0

 [B0 (EXIT)]
   Preds (1): B1
```
Previously, we gathered the topmost nullability of simple expressions by pattern matching on `DeclRefExpr` AST nodes and reading their nullability annotations. In the CFG above, block B1.1 corresponds to a `DeclRefExpr`. However, note that there is no `DeclRefExpr` directly preceding the second dereference of `x`. This means that, when we previously tried to process block B1.5, we crashed on `diagnoseDereference()` due to lack of information.

Now, we try to use annotations to initialise the nullability of any kind of pointer expression, not just `DeclRefExpr`s. This allows us to handle double dereferences.

Still, we are not yet computing the correct base (i.e., child node) type for some expressions, so not all double dereferences work as expected. In particular, further work is needed to handle double derenfereces of templated expressions. Currently, we return a null pointer when the base type is not correctly computed, and the callee does a null pointer check to avoid crashes.

PiperOrigin-RevId: 495369875