Add fail-safe to prevent test crashes in nullability verification
Tests with nested templates were crashing because we do not handle nested template arguments properly yet. In particular, we cannot use the current indexing strategy on nullability vectors when dealing with nested templates. This patch introduces a fail-safe to prevent out-of-bound crashes. Now, tests with nested templates may still fail (specifically, through false negatives), but should not crash. We will handle the false negatives in a future CL.

PiperOrigin-RevId: 496126213