Remove variant checks when they precede a `std::get<>(Variant)` call.
If the wrong variant is active, `std::get<>` will throw an exception, and terminate the process because exceptions are disabled. If we don't want to rely on exceptions terminating the process, then IMO we should use a function that is like std::get but crashes, as so:

```cc
template <size_t i, typename T>
auto GetOrDie(T&& variant) noexcept -> decltype(std::get<i>(variant)) {
  return std::get<i>(std::forward(variant));
}
```

The `assert(!IsError())` calls don't have as much information about what went wrong as the exception might, and are easy to forget. When it's an `assert`, it only ever does things in debug mode anyway.

PiperOrigin-RevId: 498611220