Make payload of `NamespaceQualifier` private.
This CL prevents direct construction of `NamespaceQualifier` and forces
using explicit `code_gen_utils` APIs to construct a
`NamespaceQualifier`.  This is a step toward a future where the provided
construction API performs *early* verification of the input (e.g.
panicking or returning an error early when the inputs cannot be
formatted as C++ namespace identifiers or Rust module names).

This CL refactors places that used to depend on a public payload
of `NamespaceQualifier` in the following way:

* Construction of `NamespaceQualifier` now goes through an explicit
  `NamespaceQualifier::new` API.

* `CratePath` has been refactored to avoid having to rip-open and
  deconstruct a `NamespaceQualifier`.

PiperOrigin-RevId: 499270950