Refactor `ImplKind::{Struct,Trait}` to store `Rc<Record>`.
This way we can remove the `record_name` and `record_qualifier` fields, calculating them inline where needed from the record itself.
This also paves the way to be able to use `Self` instead of `crate::Foo` in bindings (since for this we need to compare the `Record` of the params against the `Record` of the ImplKind inside generate_func).

PiperOrigin-RevId: 499911618