Introduce `assert_nullability`, a debugging function for nullability verification
Currently, the way in which we test the presence of nullability annotations is a bit limited. We can only check it through explicitly dereferencing pointer expressions and checking whether an "unsafe" warning is emitted. This can be inconvienent, because one needs several lines of code (i.e., several dereferences) to test a single expression that contains several pointers.

Convenience aside, our current check is also ambiguous. Firstly, it cannot distinguish between nonnull and unspecified annotations, as dereferencing either of these types is considered a safe operation. Secondly, on an unsafe dereference of a nested expression, it cannot pinpoint *which* subexpression caused the unsafety (consider, for example, a nullable pointer to a pair of nonnull pointers, or a nonnull pointer to a pair of nullable pointers; `*myPair->first` would be unsafe in both cases, but for different reasons. More granular information about which subexpression caused the nullable-ness would be helpful for debugging).

To solve these issues, we have implemented `__assert_nullability`, a special function to check, in one line of code, whether all the nullability information computed for an expression corresponds to what we expect.  All the user needs to do is pass the expression of interest as argument to `__assert_nullability`, and write the expression's expected nullability as a template parameter.

Consider:
```
    enum NullabilityKind{
      _nonnull,
      _nullable,
      _unspecified,
    };

    template<NullabilityKind ...NK, typename T>
    void __assert_nullability(T&);

    template<typename T0, typename T1>
    struct Struct2Arg{};

    void target(Struct2Arg<int *, int * _Nullable> p){
      __assert_nullability<_unspecified, _nullable>(p);
    }
```

During nullability diagnosis (in particular, `diagnoseCallExpr`), we now check whether the function being diagnosed is called `__assert_nullability`; if so, then we retrieve from the lattice the nullability vector computed for the function's argument (in the example above, `p`), and compare it with the nullabilities passed to the call as a template parameter pack (in the example, `_unspecified, _nullable`). If the two vectors are not equivalent, we signal an error and print both the expected and the computed nullability.

Future improvements:
- Add the definition of `__assert_nullability` and `NullabilityKind` as a preamble to every test, to avoid unneccesary repetition. This change will also involve adjusting the line counting offset so that line numbers output by the diagnoser reflect those seen by the user (i.e., line count should not include the preamble).
- Add another type of diagnostic check in the testing framework so we can distinguish between regular unsafe warnings and warnings of the kind "failing `__assert_nullability` check".

PiperOrigin-RevId: 500745169