Fix the broken hash implementation for Crubit strong strings.
Warnings were disabled, and so it was not a compilation error to forget to return from a function. Instead, we get undefined behavior -- in my particular case, the instruction pointer fell off the end of the function and hit the padding, which consisted of an instruction that emitted a `SIGTRAP`. Surprisingly frustrating to debug such a minor problem, but I've re-enabled warnings in a separate CL.

Huge thanks to Richard Smith and Thomas Köppe, who helped me find the bug even though I asked in the wrong channel (oops).

PiperOrigin-RevId: 501661323