Panic if `CcPrerequisites::defs` dependency cycle is detected.
In the past I have incorrectly thought that some Rust APIs cannot be
translated into C++ bindings, because of a dependency cycle that cannot
be broken by a forward declaration.  I have recently learned that the
example based on the static methods doesn't necessarily need to form a
cycle (details in b/260725687#comment3).  Therefore at this point there
are no known scenarios where `CcPrerequisites::defs` can form a cycle.

Based on the above, this CL makes the following changes:
* Asserts that `failed_ids` returned from `toposort` are always empty
* Removes formatting `failed_ids` into unsupported-item comments in the
  generated `..._cc_api.h` file.
* Trims the rationale for having a custom `toposort.rs` crate instead
  of reusing one of Cargo crates.

PiperOrigin-RevId: 501682999