C++ function declarations only need `CcPrerequsites::fwd_decls`.
A Rust function that is `#[no_mangle]` and `extern "C"` only needs
to be redeclared in the generated `..._cc_api.h` file - there is
no need to forward the call to a thunk and therefore there is no
need for a function definition.

Forward declarations are sufficient for parameter types used in function
declarations (i.e. these parameter types don't require a full definition
before the function declaration).

For the scenario described above this CL:
* Adds tests
* Fixes how `format_fn` populates `CcPrerequsites`

PiperOrigin-RevId: 501687846