Simplify handling of function calls, and prepare support for callbacks.
Function calls are now handled by creating a generic FunctionLifetimes for each
call, and imposing constraints on that FunctionLifetimes and objects that are
passed to the call.

This should be reasonably easy to extend to callbacks.

There are a couple of yet-unsolved issues with this approach, related to the
handling of flow-sensitiveness:
- Function calls are no longer a point where flow sensitiveness might cause
  variable lifetimes to change, hence this causes some overly-constraining results.
- After a function call, a local variable may no longer be overwritable; some infra
  is in place to handle this, but as of now there is no testcase that shows that
  this might happen, and the code is not complete.

PiperOrigin-RevId: 502403144