Add support for non-flow-sensitive nullability of parenthesised expressions
The CFG often omits parenthesis of expressions, but the AST does not. Therefore, the CFG match switch used to apply transfer functions does not match parenthesised expressions. This could lead to missing entries when we look up an expression's nullability in the lattice. Now, when searching for or inserting the nullability of expressions in the lattice, we skip its parenthesis.

PiperOrigin-RevId: 502635902