Add `--crubit-support-path` and `rstd::Char` type.
This CL adds `rstd::Char` - a C++ type that represents Rust's `char`
type that (a bit unlike `char32_t`) is 1) guaranteed to be 32-bits wide,
and 2) will one day reject bit patterns that lead to undefined behavior
in Rust (this is planned for a follow-up CL that should come fairly
soon).

The CL comes with only minimal tests that verify that `rstd::Char` can
be created from C++ character literals, can be explicitly cast to
`std::uint32_t`, and can be compared with itself.  The shape of the
remaining API of `rstd::Char` is not fully defined at this point (e.g.
exact behavior of the default constructor, exposing other casts,
streaming via `operator<<`, etc.).

PiperOrigin-RevId: 502682654