Detect/log/recover mismatch between expression and nullability vector
The nullability vector for an expr is supposed to match that expr's type.
If not, it's meaningless because we can't correspond entries to PointerTypes.
This requires various places that manipulate nullability vectors to be in sync,
but they very often go out of sync e.g. due to unhandled nodes.
Currently this mostly results in the contents being misinterpreted in arbitrary
ways.

Often (not always!) we can detect this as a nullability vector with the wrong
length. For now this is way too common for us to CHECK, but we can log the error
and continue by discarding all nullability info.

Logs look like:

	[ RUN      ] PointerNullabilityTest.NonNullPtrImplicitCastToBool
	=== Bad computed nullability: ===
	Expression:
	ImplicitCastExpr 0x172fbf641688 '_Bool' <PointerToBoolean>
	`-ImplicitCastExpr 0x172fbf641670 'int * _Nonnull':'int *' <LValueToRValue>
	  `-DeclRefExpr 0x172fbf641650 'int * _Nonnull':'int *' lvalue ParmVar 0x172fbf641410 'x' 'int * _Nonnull':'int *'

	Nullability (1 pointers): [_Nonnull]

	Type (0 pointers):
	BuiltinType 0x172fbfdea5d0 '_Bool'
	=================================
	=== Bad computed nullability: ===
	Expression:
	ImplicitCastExpr 0x172fbf73c908 '_Bool' <PointerToBoolean>
	`-ImplicitCastExpr 0x172fbf73c8f0 'int * _Nonnull':'int *' <LValueToRValue>
	  `-DeclRefExpr 0x172fbf73c8d0 'int * _Nonnull':'int *' lvalue ParmVar 0x172fbf73c690 'x' 'int * _Nonnull':'int *'

	Nullability (1 pointers): [_Nonnull]

	Type (0 pointers):
	BuiltinType 0x172fbf60aad0 '_Bool'
	=================================
	[       OK ] PointerNullabilityTest.NonNullPtrImplicitCastToBool (13 ms)

PiperOrigin-RevId: 502833813