Unify three type nullability traversals, so they don't get out of sync
NullabilityWalker is the source of truth for which type nodes are traversed in
which sequence, and how NullabilityKinds are associated with PointerTypes.

Not a panacea, NullabilityWalker must:
 - stay in sync with expr transfer functions
 - find the same pointers regardless of sugar (I have some further idea here)

The template substitution logic gets folded in as an optional part of
getNullabilityAnnotationsFromType as Dmitri suggested.
While here, give the callback the option to bail out (and we just traverse the
template arg sans sugar) instead of returning an invalid empty vector.
This fixes a couple of testcases.

PiperOrigin-RevId: 503951667