Move remaining runtime libraries to `crubit/support`.
This is not "obviously" correct: the current contents of crubit/support are only the runtime libraries that are present in generated code. But I don't think that's an important distinction -- rather, I think it's valuable to have all user-facing libraries in the same place.

(I would be "support"-ive of basically any directory scheme that puts them together. An alternative design, for example, is to unwrap `support/*` into `third_party/crubit`, so that people depend on `crubit:foo` instead of `crubit/support:foo`. Alternatively, we could rename `support` to `public` or `lib`, but keep it as is. Or, finally, we can just keep it as is.)

For now, moving into `support` makes it easier to find the user-facing libraries.

PiperOrigin-RevId: 504107815