Return `GeneratedItem` from all top-level `generate_*` functions, deleting `RsSnippet`.
It isn't that `RsSnippet` doesn't have any conceivable uses, it's just that `GeneratedItem` subsumes them at the top-level. We could still re-introduce an `RsSnippet`-like type to replace TokenStream at the level of expressions. The one place we were still using it is substantially simpler if it uses `GeneratedItem` instead.

(I seem to recall being unsure at the time of their introduction of whether `RsSnippet` and `GeneratedItem` were both trying to fill the same shoes. I think this CL shows they kind of were.)

The main reason for this is that I want to add things to `GeneratedItem` that control code generation. In particular, marking a generated item as experimental, supported for all rust-supporting targets, etc.  To do that, first, everything should return a `GeneratedItem`.

PiperOrigin-RevId: 504645914