Add fail-safe to prevent crashes on nullability verification
In PointerNullabilityAnalysis::diagnoseReturn, we assert that the return type of the input expression is a pointer, because this action is run on ReturnStmts of pointer type. However, a function could have a ReturnStmt of pointer type and not have a pointer return type (see test case added), which would cause the assertion to fail. In fact, we have seen this assertion fail in real files. This fix prevents this failure.

A desirable future feature would be to actually analyse non-pointer return types too.

PiperOrigin-RevId: 504809450