Represent enum bool literals as `true` or `false`, not `0` or `1`.
I removed the `ToTokens` impl for `IntegerConstant` in case we ever would be tempted to use it somewhere else -- the correct tokens depend on the type that the integer constant is being read as. (We may end up e.g. reusing this for actual non-enum constants, in particular, and would want to factor out the boolean check code at that time.) For now, since it only has one place it's used...

PiperOrigin-RevId: 506724358