Avoiding calling `FullyQualifiedName::new` for `impl` blocks.
Before this CL, `format_item` would return an error for `impl` blocks.
The errors are sorted with other snippets into the right location
and namespace which requires a call to `FullyQualifiedName::new`.
And before and after this CL `FullyQualifiedName::new` panics
when trying to get a `Symbol` associated with an `impl` block.

After this CL, `format_item` returns `None` to silently skip
`impl` blocks.  This is only temporary - in the next CL we will
start to format contents of the `impl` blocks (initially just
emitting errors for all individual impl items).

PiperOrigin-RevId: 508753790