Introduce separate `SnippetKey` and `SnippetKind` types.
This is the 1st CL out of a series of 3 CLs that work toward returning
separate snippets from `format_fn` for A) declaration of the main API
function, and B) thunk declaration, function definition, and other
implementation details.  After the 3 CLs land, adding support for
static methods should be fairly straightforward.

PiperOrigin-RevId: 508770737