Return multiple snippets from `format_adt`.
This is the 2nd CL out of a series of 3 CLs that work toward returning
separate snippets from `format_fn` for A) declaration of the main API
function, and B) thunk declaration, function definition, and other
implementation details.  After the 3 CLs land, adding support for
static methods should be fairly straightforward.

This CL also has a side effect of moving C++ assertions to the end of
the generated header (grouping public APIs together, at the top of
the header).

PiperOrigin-RevId: 508774419