Delete `IR::is_stdlib_target()`.
I wanted to delete this to see what goes wrong, so that I could see why it was here. Instead, no tests broke. So I guess it doesn't have a reason to be here -- maybe it used to before we moved the standard library into its own (not "fake") BUILD rule?

PiperOrigin-RevId: 509518224