Propagate `--crate-name` into `cc_bindings_from_rs`.
This CL is needed to support running `cc_bindings_from_rs` on any Rust
crate where the name of the crate is not based on the basename of the
crate root (the name of the main `.rs` file).

This CL can be seen as a *small* step toward replicating all `rustc`
cmdline arguments into the invocation of `cc_bindings_from_rs`.

This CL makes the following changes:

* Adds `--crate-name` propagation to
  `cc_bindings_from_rs/bazel_support/cc_bindings_from_rust_rule.bzl`
* Adds tests that exercise the `--crate-name` support and would fail
  without the change above: `cc_bindings_from_rs/test/bazel/crate_name`
* Moves tests from `cc_bindings_from_rs/test/bazel_unit_tests` to
  `cc_bindings_from_rs/test/bazel/unit_tests`.  This helps to keep
  the Bazel-focused tests under `cc_bindings_from_rs/test/bazel`.

PiperOrigin-RevId: 510499869