Rename `rstd` to `rs_std`.
We are still discussing whether in the long-term we want to have both
`rs_std` (with C++ bindings for Rust standard library) and `rs_ffi`
(with C++ helpers for working with Rust language and its builtin types
like `char`).  But regardless of the outcome of this discussion, there
is a `rs_std` C++ library and namespace in the future, so renaming
`rstd` to `rs_std` seems like a step in the right direction.  And until
this discussion reaches actionable conclusions, it seems fine to keep
`Char` (or `rs_char` or `rchar` after renaming) in the `rs_std`
namespace.

PiperOrigin-RevId: 510522956