Don't generate bindings for the `rs_std::rs_char` struct.
This CL disables generation of bindings for the `rs_std::rs_char` type
from `crubit/support/rs_std/rs_char.h`.

This CL is compatible with round-tripping `rs_char` into the built-in
`char` type in the future.  (Although in that case bindings for the
`rs_char` struct should be *skipped* rather than resulting in a
comment pointing it out as an unsupported item.)

It is conceivable that we may decide to revert this CL in the future
(and provide `impl Into<char> for rs_std::rs_char`).  In the meantime,
avoiding generating the bindings seems to keep all our options open.

PiperOrigin-RevId: 511332771